requirements.txt
# qm_final3 系统依赖

# 基础科学计算
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0

# 深度学习框架
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0

# 音视频处理
opencv-python>=4.5.0
librosa>=0.9.0
soundfile>=0.10.0
pyaudio>=0.2.11

# Web框架和API
gradio>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# 多模态和NLP
transformers>=4.30.0
mediapipe>=0.10.0
openai-whisper>=20230314

# 数据处理和存储
pandas>=1.3.0
pyyaml>=6.0
redis>=4.0.0

# 系统监控
psutil>=5.8.0
gputil>=1.4.0

# 异步处理
asyncio-mqtt>=0.11.0
aiofiles>=23.0.0

# 工具库
tqdm>=4.62.0
pillow>=9.0.0
click>=8.0.0
pydantic>=2.0.0

# 开发和测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# 可选依赖（用于高级功能）
# 音乐生成
# musicgen>=1.0.0

# 图神经网络（用于KG-MLP）
# torch-geometric>=2.3.0
# dgl>=1.1.0

# 知识图谱
# rdflib>=6.0.0
# neo4j>=5.0.0